Rename MetricsCard filename.
